Add tests to backend (#7)

+++ judge-backend/common/src/main/kotlin/io/github/leonfoliveira/judge/common/adapter/oauth/OAuthJwtAdapter.kt
@@ -27,9 +27,9 @@ class OAuthJwtAdapter(
         val expiresAt =
             now.plusMillis(
                 if (member.type == Member.Type.ROOT) {
-                    UnitUtil.parseTimeValue(rootExpiration).toLong()
+                    UnitUtil.parseTimeValue(rootExpiration)
                 } else {
-                    UnitUtil.parseTimeValue(expiration).toLong()
+                    UnitUtil.parseTimeValue(expiration)
                 },
             )
